Separate i18n code into multiple files to fix circular dependencies
